requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
xgboost
catboost
python-dotenv
pymongo[srv]==4.1.1
certifi
PyYAML
-e .